Add support for comment events for open deployments
Extract a separate module from RealtimeHapiPlugin
