+++ src/realtime/realtime-hapi-plugin.ts
@@ -1,99 +1,37 @@
-import { Observable, Subscription } from '@reactivex/rxjs';
+import { Observable } from '@reactivex/rxjs';
+import * as Boom from 'boom';
 import { inject, injectable } from 'inversify';
 import * as Joi from 'joi';
-import { isNil, omitBy } from 'lodash';
 import * as moment from 'moment';
 
 import {
-  createActivityEvent,
-  MinardActivity,
-} from '../activity';
-import {
   COMMENT_ADDED_EVENT_TYPE,
   COMMENT_DELETED_EVENT_TYPE,
-  CommentAddedEvent,
-  CommentDeletedEvent,
-  createCommentAddedEvent,
-  createCommentDeletedEvent,
 } from '../comment';
-import {
-  createDeploymentEvent,
-  DeploymentEvent,
-} from '../deployment';
-import { eventBusInjectSymbol, PersistentEventBus } from '../event-bus/';
-import { ApiProject, JsonApiHapiPlugin } from '../json-api';
-import {
-  ApiEntity,
-  toApiBranchId,
-  toApiCommitId,
-  toApiDeploymentId,
-} from '../json-api';
-import {
-  codePushed,
-  CodePushedEvent,
-  projectCreated,
-  ProjectCreatedEvent,
-  projectDeleted,
-  ProjectDeletedEvent,
-  projectEdited,
-  ProjectEditedEvent,
-} from '../project';
 import * as Hapi from '../server/hapi';
 import { HapiPlugin } from '../server/hapi-register';
-import {
-  Event,
-  eventCreator,
-  isPersistedEvent,
-  isType,
-  PersistedEvent,
-  StreamingEvent,
-} from '../shared/events';
+import { PersistedEvent } from '../shared/events';
 import * as logger from '../shared/logger';
 import { ObservableWrapper } from './observable-wrapper';
-import {
-  StreamingCodePushedEvent,
-  StreamingCommentDeletedEvent,
-  StreamingDeploymentEvent,
-} from './types';
+import { RealtimeModule } from './realtime-module';
 
 export const PING_INTERVAL = 20000;
 
-type StreamFactory = (request: Hapi.Request) => Promise<Observable<PersistedEvent<any>>>;
+type Predicate = (event: PersistedEvent<any>) => boolean;
 
 @injectable()
 export class RealtimeHapiPlugin extends HapiPlugin {
 
   public static injectSymbol = Symbol('realtime-plugin');
-  private eventBusSubscription: Subscription;
-  public readonly persistedEvents: Observable<PersistedEvent<any>>;
 
   constructor(
-    @inject(JsonApiHapiPlugin.injectSymbol) private readonly jsonApiPlugin: JsonApiHapiPlugin,
-    @inject(eventBusInjectSymbol) private readonly eventBus: PersistentEventBus,
+    @inject(RealtimeModule.injectSymbol) private readonly realtimeModule: RealtimeModule,
     @inject(logger.loggerInjectSymbol) private readonly logger: logger.Logger,
   ) {
     super({
-        name: 'realtime-plugin',
-        version: '1.0.0',
+      name: 'realtime-plugin',
+      version: '1.0.0',
     });
-    this.persistedEvents = this.eventBus.getStream()
-      .filter(isPersistedEvent)
-      .map(event => event as PersistedEvent<any>)
-      .share();
-
-    // creates SSEEvents and posts them
-    this.eventBusSubscription = this.getEnrichedStream()
-      .subscribe(this.eventBus.post.bind(this.eventBus));
-
-    this.getStream = this.getStream.bind(this);
-  }
-
-  private getEnrichedStream(): Observable<StreamingEvent<any>> {
-    return this.enrich(this.eventBus.getStream())
-      .catch(err => {
-        this.logger.error('Error on enrich:', err);
-        return this.getEnrichedStream();
-      });
   }
 
   public register(server: Hapi.Server, _options: Hapi.IServerOptions, next: () => void) {
@@ -102,7 +40,7 @@ export class RealtimeHapiPlugin extends HapiPlugin {
       method: 'GET',
       path: '/events/{teamId}',
       handler: {
-        async: this.requestHandlerFactory(this.getStream),
+        async: this.teamHandler,
       },
       config: {
         bind: this,
@@ -116,13 +54,13 @@ export class RealtimeHapiPlugin extends HapiPlugin {
       },
     }, {
       method: 'GET',
-      path: '/events/{teamId}/deployment/{projectId}-{deploymentId}',
+      path: '/events/deployment/{projectId}-{deploymentId}',
       handler: {
-        async: this.requestHandlerFactory(this.getStream),
+        async: this.deploymentHandler,
       },
       config: {
         bind: this,
-        auth: 'jwt-url',
+        auth: false,
         cors: true,
         validate: {
           params: {
@@ -136,204 +74,65 @@ export class RealtimeHapiPlugin extends HapiPlugin {
     next();
   }
 
-  private async getStream(request: Hapi.Request) {
-    const {teamId, projectId, deploymentId} = request.params;
-    const _teamId = parseInt(teamId, 10);
-    const since = parseInt(request.headers['last-event-id'], 10);
-    let predicate = (event: PersistedEvent<any>) => event.teamId === _teamId;
-    if (projectId && deploymentId) {
-      const _projectId = parseInt(projectId, 10);
-      const _deploymentId = parseInt(deploymentId, 10);
-      predicate = deploymentEventFilter(_teamId, _projectId, _deploymentId);
-    }
-
-    let stream = Observable.concat(
+  public async getStream(teamId: number, predicate: Predicate, lastEventId?: number) {
+    const stream = Observable.concat(
+      lastEventId ?
+        await this.realtimeModule.getExistingEvents(teamId, lastEventId) :
+        Observable.empty<PersistedEvent<any>>(),
+      this.realtimeModule.getSSEStream(),
+    );
+    return Observable.concat(
       Observable.of(pingEventCreator()),
-      this.persistedEvents,
-    );
-
-    if (since && !isNaN(since)) {
-      const existing = await this.eventBus.getEvents(_teamId, since);
-      if (existing.length > 0) {
-        existing.shift(); // getEvents is '>= since', but here we want '> since'
-      }
-      stream = Observable.concat(Observable.from(existing), stream);
-    }
-
-    return Observable.merge(
-      Observable.interval(PING_INTERVAL).map(_ => pingEventCreator()),
-      stream.filter(predicate),
+      Observable.merge(
+        Observable.interval(PING_INTERVAL).map(_ => pingEventCreator()),
+        stream.filter(predicate),
+      ),
     );
   }
 
-  private requestHandlerFactory(streamFactory: StreamFactory) {
-    return async (request: Hapi.Request, reply: Hapi.IReply) => {
-      try {
-        // const observable = await this.getObservable(teamId, since ? parseInt(since, 10) : undefined);
-        const observable = await streamFactory(request);
-        const nodeStream = new ObservableWrapper(observable);
-        reply(nodeStream)
-          .header('content-type', 'text/event-stream')
-          .header('content-encoding', 'identity');
+  private async streamReply(stream: Observable<PersistedEvent<any>>, request: Hapi.Request, reply: Hapi.IReply) {
+    const nodeStream = new ObservableWrapper(stream);
+    reply(nodeStream)
+      .header('content-type', 'text/event-stream')
+      .header('content-encoding', 'identity');
 
-        request.once('disconnect', () => {
-          // Clean up on disconnect
-          nodeStream.push(null);
-        });
-
-      } catch (err) {
-        this.logger.error('Error handling a SSE request', err);
-      }
-    };
+    request.once('disconnect', () => {
+      // Clean up on disconnect
+      nodeStream.push(null);
+    });
   }
 
-  private enrich(stream: Observable<Event<any>>): Observable<StreamingEvent<any>> {
-    return stream
-      .flatMap(event => {
-        if (isType<ProjectCreatedEvent>(event, projectCreated)) {
-          return this.projectCreated(event);
-        }
-
-        if (isType<ProjectDeletedEvent>(event, projectDeleted)) {
-          return Observable.of(this.toSSE(event, event.payload));
-        }
-
-        if (isType<ProjectEditedEvent>(event, projectEdited)) {
-          return this.projectEdited(event);
-        }
-
-        if (isType<CodePushedEvent>(event, codePushed)) {
-          return this.codePushed(event);
-        }
-
-        if (isType<MinardActivity>(event, createActivityEvent)) {
-          return this.activity(event);
-        }
-
-        if (isType<DeploymentEvent>(event, createDeploymentEvent)) {
-          return this.deployment(event);
-        }
-
-        if (isType<CommentAddedEvent>(event, createCommentAddedEvent)) {
-          return this.commentAdded(event);
-        }
-
-        if (isType<CommentDeletedEvent>(event, createCommentDeletedEvent)) {
-          return this.commentDeleted(event);
-        }
-
-        return Observable.empty<StreamingEvent<any>>();
-      }, 3);
-  }
-
-  private async projectEdited(event: Event<ProjectEditedEvent>) {
-    const payload = omitBy({
-      id: event.payload.id,
-      name: event.payload.name,
-      description: event.payload.description,
-      'repo-url': event.payload.repoUrl,
-    }, isNil);
-    return this.toSSE(event, payload);
-  }
-
-  private async activity(event: Event<MinardActivity>) {
-    const apiActivity = await this.jsonApiPlugin.getJsonApiModule().toApiActivity(event.payload);
-    const payload = this.jsonApiPlugin.serializeApiEntity('activity', apiActivity).data;
-    return this.toSSE(event, payload);
-  }
-
-  private async codePushed(event: Event<CodePushedEvent>) {
+  private async teamHandler(request: Hapi.Request, reply: Hapi.IReply) {
     try {
-      const projectId = event.payload.projectId;
-      const commits = await Promise.all(event.payload.commits.map(async item => {
-        const apiCommit = await this.getJsonApiModule().toApiCommit(projectId, item, []);
-        return this.serializeApiEntity('commit', apiCommit).data;
-      }));
-
-      const branch = event.payload.after ?
-        (await this.jsonApiPlugin.getEntity('branch', api => api.getBranch(projectId, event.payload.ref))).data :
-        toApiBranchId(projectId, event.payload.ref);
-
-      const payload: StreamingCodePushedEvent = {
-        teamId: event.payload.teamId,
-        commits,
-        after: event.payload.after ? toApiCommitId(projectId, event.payload.after.id) : undefined,
-        before: event.payload.before ? toApiCommitId(projectId, event.payload.before.id) : undefined,
-        parents: event.payload.parents.map(item => toApiCommitId(projectId, item.id)),
-        branch,
-        project: String(projectId),
-      };
-      return this.toSSE(event, payload);
-
-    } catch (error) {
-      this.logger.error(`Unable to create StreamingCodePushedEvent`, { error, event });
-      throw Error('Unable to create StreamingCodePushedEvent');
+      const { teamId } = request.params;
+      const _teamId = parseInt(teamId, 10);
+      const predicate = (event: PersistedEvent<any>) => event.teamId === _teamId;
+      const stream = await this.getStream(_teamId, predicate, getLastEventId(request));
+      return this.streamReply(stream, request, reply);
+    } catch (err) {
+      this.logger.warn('Problems handling a SSE request', err);
+      return reply(Boom.wrap(err));
     }
   }
 
-  private getJsonApiModule() {
-    return this.jsonApiPlugin.getJsonApiModule();
-  }
-
-  private serializeApiEntity(type: string, entity: ApiEntity) {
-    return this.jsonApiPlugin.serializeApiEntity(type, entity);
-  }
-
-  private async commentAdded(event: Event<CommentAddedEvent>) {
-    const comment = await this.getJsonApiModule().toApiComment(event.payload);
-    const payload = this.serializeApiEntity('comment', comment).data;
-    return this.toSSE(event, payload);
-  }
-
-  private async commentDeleted(event: Event<CommentDeletedEvent>) {
-    const payload: StreamingCommentDeletedEvent = {
-      comment: String(event.payload.commentId),
-      teamId: event.payload.teamId,
-      deployment: toApiDeploymentId(event.payload.projectId, event.payload.deploymentId),
-    };
-    return this.toSSE(event, payload);
-  }
-
-  private async projectCreated(event: Event<ProjectCreatedEvent>) {
-    const payload: ApiProject = await this.jsonApiPlugin
-      .getEntity('project', api => api.getProject(event.payload.id));
-    return this.toSSE(event, payload);
-  }
-
-  private async deployment(event: Event<DeploymentEvent>) {
+  private async deploymentHandler(request: Hapi.Request, reply: Hapi.IReply) {
     try {
-      const deployment = event.payload.deployment;
-      const apiResponse = await this.jsonApiPlugin.getEntity(
-        'deployment', api => api.toApiDeployment(deployment.projectId, deployment));
-
-      const ssePayload: StreamingDeploymentEvent = {
-        teamId: event.payload.teamId,
-        branch: toApiBranchId(deployment.projectId, deployment.ref),
-        project: String(deployment.projectId),
-        commit: toApiCommitId(deployment.projectId, deployment.commitHash),
-        deployment: apiResponse.data,
-      };
-      return this.toSSE(event, ssePayload);
-    } catch (error) {
-      const msg = 'Could not convert DeploymentEvent to streaming event';
-      this.logger.error(msg, { error, event });
-      throw Error(msg);
-    }
-  }
-
-  private toSSE<T>(event: Event<any>, payload: T): StreamingEvent<T> {
-    if (typeof event.teamId !== 'number' && typeof (payload as any).teamId !== 'number') {
-      throw Error('Tried to convert an incompatible event to an SSEEvent');
-    }
-
-    const type = `SSE_${event.type}`;
-    return eventCreator<any>(type, (_event: Event<any>) => {
-      if (event.teamId) {
-        _event.teamId = event.teamId!;
+      const { projectId, deploymentId } = request.params;
+      const teamId = await request.getOpenTeamId();
+      if (projectId && deploymentId && teamId) {
+        const _projectId = parseInt(projectId, 10);
+        const _deploymentId = parseInt(deploymentId, 10);
+        const predicate = deploymentEventFilter(teamId, _projectId, _deploymentId);
+        const stream = await this.getStream(teamId, predicate, getLastEventId(request));
+        return this.streamReply(stream, request, reply);
       }
-      return true;
-    })(payload) as StreamingEvent<T>;
+    } catch (err) {
+      this.logger.warn('Problems handling a SSE request', err);
+      return reply(Boom.wrap(err));
+    }
+    return reply(Boom.notFound());
   }
+
 }
 
 export function pingEventCreator(): PersistedEvent<any> {
@@ -359,4 +158,12 @@ export function deploymentEventFilter(teamId: number, projectId: number, deploym
     }
     return false;
   };
+}
+
+export function getLastEventId(request: Hapi.Request) {
+  const since = parseInt(request.headers['last-event-id'], 10);
+  if (since && !isNaN(since)) {
+    return since;
+  }
+  return undefined;
 }
